Prepare for heroku deployment.
Changes based on https://github.com/mars/create-react-app-buildpack#customization

Observe the following citation from the documentation:

```
Warning: If replacing the default web process, please check this
buildpack's Purpose to avoid misusing this buildpack (such as running
a Node server) which can lead to confusing deployment issues.
```

The warning here seems very relevant to the changes that we have
introduced. But we haven't (at least not yet) created a separate node
package with the node server so the different other recipes mentioned
as a better alternative did not seem relevant to us at the time. See
https://github.com/mars/heroku-cra-node