Main tag should be placed outside the form tag
Moved main to wrap page title and NavForm, and form deeper, to only wrap the "wizard-page" component.
Unsure if main should contain the StepIndicator, but we will test this in a few days and move it if it becomes annoying to navigate past that section every time we change steps.