Make scripts cross-platform
First, this commit adds cross-env, to convert use (read and write) of environment variables to cross-platform.
Second, it adds DECORATOR_URL to "npm start" in server/package.json (also removing this from "dev-server" script in /package.json.
Third, it removes use of "--prefix" syntax, as it is unnecessary and didn't work on windows.

Finally, on windows, you have to switch default npm shell to Bash in order for backtick syntax to work.
Check .npmrc in your user's home folder, and add 'script-shell="path/to/bash" '.