Submit/go to next step when pressing enter in form.
This will push the buttons into the blue frame around forms, which is not ideal, but required to get this to work for now.
Adding form-attribute to buttons outside of the form does not work, as form.io seem to filter that attribute away.
Adding a div inside the form for styling, but around the components somehow breaks validation, so that was not an option either.