Override Form.io checkCondition function to handle chained lookups in submissionData.
Checks for "a.b.c" in submissionData is now automatically replaced by "a && a.b && a.b.c" to make it easier for editors/form owners to make custom conditionals.
This relates specifically to custom conditionals depending on a conditional "Flervalg" component, which does not exist in submissionData until filled out.
Default value is "show = true" when an error occurs, meaning conditionally rendered components in these cases will always be rendered.