Add type="button" to all buttons that should not submit.
Some of these may be unneccessary, but they won't break anything.
However, a button without type, which is placed inside a form, will be interpreted as a submit-button.
As a result, pressing the enter key anywhere inside a form will activate the first such button.
An example is while adding options to a RadioPanelGruppe, where pressing enter in the label input would add a new option. This change fixes this.