NavDatepicker submission is cleared when navigating to SummaryPage.
It is not quite clear *why* this code causes that, but it probably has something to do with trying to update a form that has been removed, and the React components will then be hidden.
Maybe the others "live on" in the Formio-object? Anyhow, this fixes the problem, and doesn't appear to cause any new bugs.