Use useCallback to avoid re-fetching translations (and countries-list) each time the URL/path changes.
This didn't use to be a huge problem, but we fetched translations and countries-list again when we got to the summary page, and subsequent URLs.
However, with the new routing task on the way, we will have unique URLs for each step in the form, causing re-fetching, and potential re-rendering, on each step.
This could even cause loss of data.

This task uses useCallback to memoize functions from useTranslations, so they don't change between re-renders.