Adds custom command for clicking send digital link
Introduces a custom Cypress command, `clickSendDigital`, to streamline the process of clicking the "Send Digital" link on the introduction page.
This simplifies tests and reduces code duplication by encapsulating the logic for finding and clicking the correct link, which can vary based on login state.
Also refactors tests to use this new command, enhancing readability and maintainability.